+++ pypsrp/client.py
@@ -5,19 +5,19 @@ from __future__ import division
 
 import base64
 import hashlib
+import logging
 import os
 import shutil
 import sys
 import tempfile
-
-import logging
+import warnings
 
 from pypsrp.exceptions import WinRMError
-from pypsrp.powershell import PowerShell, RunspacePool
+from pypsrp.powershell import DEFAULT_CONFIGURATION_NAME, PowerShell, RunspacePool
 from pypsrp.serializer import Serializer
 from pypsrp.shell import Process, SignalCode, WinRS
 from pypsrp.wsman import WSMan
-from pypsrp._utils import to_bytes, to_string, to_unicode
+from pypsrp._utils import get_pwsh_script, to_bytes, to_unicode
 
 if sys.version_info[0] == 2 and sys.version_info[1] < 7:  # pragma: no cover
     # ElementTree in Python 2.6 does not support namespaces so we need to use
@@ -51,7 +51,7 @@ class Client(object):
         """
         self.wsman = WSMan(server, **kwargs)
 
-    def copy(self, src, dest):
+    def copy(self, src, dest, configuration_name=DEFAULT_CONFIGURATION_NAME):
         """
         Copies a single file from the current host to the remote Windows host.
         This can be quite slow when it comes to large files due to the
@@ -65,9 +65,11 @@ class Client(object):
 
         :param src: The path to the local file
         :param dest: The path to the destionation file on the Windows host
+        :param configuration_name: The PowerShell configuration endpoint to
+            use when copying the file.
         :return: The absolute path of the file on the Windows host
         """
-        def read_buffer(b_path, buffer_size):
+        def read_buffer(b_path, total_size, buffer_size):
             offset = 0
             sha1 = hashlib.sha1()
 
@@ -77,18 +79,17 @@ class Client(object):
                               % (offset, buffer_size))
                     offset += len(data)
                     sha1.update(data)
-                    b64_data = base64.b64encode(data) + b"\r\n"
+                    b64_data = base64.b64encode(data)
 
-                    yield b64_data, False
+                    result = [to_unicode(b64_data)]
+                    if offset == total_size:
+                        result.append(to_unicode(base64.b64encode(to_bytes(sha1.hexdigest()))))
+
+                    yield result
 
                 # the file was empty, return empty buffer
                 if offset == 0:
-                    yield b"", False
-
-            # the last input is the actual file hash used to verify the
-            # transfer was ok
-            actual_hash = b"\x00\xffHash: " + to_bytes(sha1.hexdigest())
-            yield base64.b64encode(actual_hash), True
+                    yield [u"", to_unicode(base64.b64encode(to_bytes(sha1.hexdigest())))]
 
         src = os.path.expanduser(os.path.expandvars(src))
         b_src = to_bytes(src)
@@ -96,82 +97,34 @@ class Client(object):
         log.info("Copying '%s' to '%s' with a total size of %d"
                  % (src, dest, src_size))
 
-        # check if the src size is twice as large as the max payload and fetch
-        # the max size from the server, we only check in this case to save on a
-        # round trip if the file is small enough to fit in 2 msg's, otherwise
-        # we want to get the largest size possible
-        buffer_size = int(self.wsman.max_payload_size / 4 * 3)
-        if src_size > (buffer_size * 2) and \
-                self.wsman.max_envelope_size == 153600:
-            log.debug("Updating the max WSMan envelope size")
-            self.wsman.update_max_payload_size()
-            buffer_size = int(self.wsman.max_payload_size / 4 * 3)
-        log.info("Creating file reader with a buffer size of %d" % buffer_size)
-        read_gen = read_buffer(b_src, buffer_size)
+        with RunspacePool(self.wsman, configuration_name=configuration_name) as pool:
+            # Get the buffer size of each fragment to send, subtract. Adjust to size of the base64 encoded bytes. Also
+            # subtract 82 for the fragment, message, and other header info that PSRP adds.
+            buffer_size = int((self.wsman.max_payload_size - 82) / 4 * 3)
 
-        command = u'''begin {
-    $ErrorActionPreference = "Stop"
-    $path = [System.IO.Path]::GetTempFileName()
-    $fd = [System.IO.File]::Create($path)
-    $algo = [System.Security.Cryptography.SHA1CryptoServiceProvider]::Create()
-    $bytes = @()
-    $expected_hash = ""
-} process {
-    $base64_string = $input
+            log.info("Creating file reader with a buffer size of %d" % buffer_size)
+            read_gen = read_buffer(b_src, src_size, buffer_size)
 
-    $bytes = [System.Convert]::FromBase64String($base64_string)
-    if ($bytes.Count -eq 48 -and $bytes[0] -eq 0 -and $bytes[1] -eq 255) {
-        $hash_bytes = $bytes[-40..-1]
-        $expected_hash = [System.Text.Encoding]::UTF8.GetString($hash_bytes)
-    } else {
-        $algo.TransformBlock($bytes, 0, $bytes.Length, $bytes, 0) > $null
-        $fd.Write($bytes, 0, $bytes.Length)
-    }
-} end {
-    $output_path = $ExecutionContext.SessionState.Path.GetUnresolvedProviderPathFromPSPath('%s')
-    $dest = New-Object -TypeName System.IO.FileInfo -ArgumentList $output_path
-    $fd.Close()
-
-    try {
-        $algo.TransformFinalBlock($bytes, 0, 0) > $null
-        $actual_hash = [System.BitConverter]::ToString($algo.Hash)
-        $actual_hash = $actual_hash.Replace("-", "").ToLowerInvariant()
-
-        if ($actual_hash -ne $expected_hash) {
-            $msg = "Transport failure, hash mistmatch"
-            $msg += "`r`nActual: $actual_hash"
-            $msg += "`r`nExpected: $expected_hash"
-            throw $msg
-        }
-        [System.IO.File]::Copy($path, $output_path, $true)
-        $dest.FullName
-    } finally {
-        [System.IO.File]::Delete($path)
-    }
-}''' % to_unicode(dest)
-        encoded_command = to_string(base64.b64encode(to_bytes(command,
-                                                              'utf-16-le')))
-
-        with WinRS(self.wsman) as shell:
-            process = Process(shell, "powershell.exe",
-                              ["-NoProfile", "-NonInteractive",
-                               "-EncodedCommand", encoded_command])
-            process.begin_invoke()
+            command = get_pwsh_script('copy.ps1')
             log.debug("Starting to send file data to remote process")
-            for input_data, end in read_gen:
-                process.send(input_data, end)
+            powershell = PowerShell(pool)
+            powershell.add_script(command).add_argument(dest)
+            powershell.invoke(input=read_gen)
             log.debug("Finished sending file data to remote process")
-            process.end_invoke()
 
-        stderr = self.sanitise_clixml(process.stderr)
-        if process.rc != 0:
-            raise WinRMError("Failed to copy file: %s" % stderr)
-        output_file = to_unicode(process.stdout).strip()
-        log.info("Completed file transfer of '%s' to '%s'"
-                 % (src, output_file))
+        for warning in powershell.streams.warning:
+            warnings.warn(str(warning))
+
+        if powershell.had_errors:
+            errors = powershell.streams.error
+            error = "\n".join([str(err) for err in errors])
+            raise WinRMError("Failed to copy file: %s" % error)
+
+        output_file = to_unicode(powershell.output[-1]).strip()
+        log.info("Completed file transfer of '%s' to '%s'" % (src, output_file))
         return output_file
 
-    def execute_cmd(self, command, encoding='437'):
+    def execute_cmd(self, command, encoding='437', environment=None):
         """
         Executes a command in a cmd shell and returns the stdout/stderr/rc of
         that process. This uses the raw WinRS layer and can be used to execute
@@ -182,6 +135,8 @@ class Client(object):
             correlates to the codepage of the host and traditionally en-US
             is 437. This probably doesn't need to be modified unless you are
             running a different codepage on your host
+        :param environment: A dictionary containing environment keys and
+            values to set on the executing process.
         :return: A tuple of
             stdout: A unicode string of the stdout
             stderr: A unicode string of the stderr
@@ -192,7 +147,7 @@ class Client(object):
         set
         """
         log.info("Executing cmd process '%s'" % command)
-        with WinRS(self.wsman) as shell:
+        with WinRS(self.wsman, environment=environment) as shell:
             process = Process(shell, command)
             process.invoke()
             process.signal(SignalCode.CTRL_C)
@@ -200,7 +155,8 @@ class Client(object):
         return to_unicode(process.stdout, encoding), \
             to_unicode(process.stderr, encoding), process.rc
 
-    def execute_ps(self, script):
+    def execute_ps(self, script, configuration_name=DEFAULT_CONFIGURATION_NAME,
+                   environment=None):
         """
         Executes a PowerShell script in a PowerShell runspace pool. This uses
         the PSRP layer and is designed to run a PowerShell script and not a
@@ -213,6 +169,10 @@ class Client(object):
         stdout/stderr/rc, use execute_cmd instead.
 
         :param script: The PowerShell script to run
+        :param configuration_name: The PowerShell configuration endpoint to
+            use when executing the script.
+        :param environment: A dictionary containing environment keys and
+            values to set on the executing script.
         :return: A tuple of
             output: A unicode string of the output stream
             streams: pypsrp.powershell.PSDataStreams containing the other
@@ -221,8 +181,19 @@ class Client(object):
                 during execution
         """
         log.info("Executing PowerShell script '%s'" % script)
-        with RunspacePool(self.wsman) as pool:
+        with RunspacePool(self.wsman, configuration_name=configuration_name) as pool:
             powershell = PowerShell(pool)
+
+            if environment:
+                for env_key, env_value in environment.items():
+                    # Done like this for easier testing, preserves the param order
+                    log.debug("Setting env var '%s' on PS script execution" % env_key)
+                    powershell.add_cmdlet('New-Item'). \
+                        add_parameter('Path', 'env:'). \
+                        add_parameter('Name', env_key). \
+                        add_parameter('Value', env_value). \
+                        add_parameter('Force', True). \
+                        add_cmdlet('Out-Null').add_statement()
 
             # so the client executes a powershell script and doesn't need to
             # deal with complex PS objects, we run the script in
@@ -237,7 +208,7 @@ class Client(object):
         return "\n".join(powershell.output), powershell.streams, \
                powershell.had_errors
 
-    def fetch(self, src, dest):
+    def fetch(self, src, dest, configuration_name=DEFAULT_CONFIGURATION_NAME):
         """
         Will fetch a single file from the remote Windows host and create a
         local copy. Like copy(), this can be slow when it comes to fetching
@@ -248,56 +219,17 @@ class Client(object):
 
         :param src: The path to the file on the remote host to fetch
         :param dest: The path on the localhost host to store the file as
+        :param configuration_name: The PowerShell configuration endpoint to
+            use when fetching the file.
         """
         dest = os.path.expanduser(os.path.expandvars(dest))
         log.info("Fetching '%s' to '%s'" % (src, dest))
 
-        self.wsman.update_max_payload_size()
-
-        # Need to output as a base64 string as PS Runspaces will create an
-        # individual byte objects for each byte in a byte array which has way
-        # more overhead than a single base64 string.
-        # I also wanted to output in chunks and have the local side process
-        # the output in parallel for large files but it seems like the base64
-        # stream is getting sent in one chunk when in a loop so scratch that
-        # idea
-        script = '''$ErrorActionPreference = 'Stop'
-$src_path = $ExecutionContext.SessionState.Path.GetUnresolvedProviderPathFromPSPath('%s')
-if (Test-Path -LiteralPath $src_path -PathType Container) {
-    throw "The path at '$src_path' is a directory, src must be a file"
-} elseif (-not (Test-Path -LiteralPath $src_path)) {
-    throw "The path at '$src_path' does not exist"
-}
-
-$algo = [System.Security.Cryptography.SHA1CryptoServiceProvider]::Create()
-$src = New-Object -TypeName System.IO.FileInfo -ArgumentList $src_path
-$buffer_size = 4096
-$offset = 0
-$fs = $src.OpenRead()
-$total_bytes = $fs.Length
-$bytes_to_read = $total_bytes - $offset
-try {
-    while ($bytes_to_read -ne 0) {
-        $bytes = New-Object -TypeName byte[] -ArgumentList $bytes_to_read
-        $read = $fs.Read($bytes, $offset, $bytes_to_read)
-
-        Write-Output -InputObject ([System.Convert]::ToBase64String($bytes))
-        $bytes_to_read -= $read
-        $offset += $read
-
-        $algo.TransformBlock($bytes, 0, $bytes.Length, $bytes, 0) > $null
-    }
-} finally {
-    $fs.Dispose()
-}
-
-$algo.TransformFinalBlock($bytes, 0, 0) > $Null
-$hash = [System.BitConverter]::ToString($algo.Hash)
-$hash.Replace("-", "").ToLowerInvariant()''' % src
-
-        with RunspacePool(self.wsman) as pool:
+        with RunspacePool(self.wsman, configuration_name=configuration_name) as pool:
+            script = get_pwsh_script('fetch.ps1')
             powershell = PowerShell(pool)
-            powershell.add_script(script)
+            powershell.add_script(script).add_argument(src)
+
             log.debug("Starting remote process to output file data")
             powershell.invoke()
             log.debug("Finished remote process to output file data")
@@ -341,14 +273,16 @@ $hash.Replace("-", "").ToLowerInvariant()''' % src
         :return: A unicode code string of the decoded output
         """
         output = to_unicode(clixml)
-        if output.startswith("#< CLIXML\r\n"):
-            serializer = Serializer()
-            output = output[11:]
+        if output.startswith("#< CLIXML"):
+            # Strip off the '#< CLIXML\r\n' by finding the 2nd index of '<'
+            output = output[clixml.index('<', 2):]
             element = ET.fromstring(output)
             namespace = element.tag.replace("Objs", "")[1:-1]
+
             errors = []
             for error in element.findall("{%s}S[@S='Error']" % namespace):
                 errors.append(error.text)
-            output = serializer._deserialize_string("".join(errors))
+
+            output = Serializer()._deserialize_string("".join(errors))
 
         return output