+++ tests/test_client.py
@@ -11,6 +11,11 @@ from pypsrp.wsman import WSMan
 from pypsrp._utils import to_unicode
 
 try:
+    from collections import OrderedDict
+except ImportError:  # pragma: no cover
+    from ordereddict import OrderedDict
+
+try:
     from unittest.mock import MagicMock
 except ImportError:
     from mock import MagicMock
@@ -82,60 +87,6 @@ class TestClient(object):
                              # checks as to whether the correct number of calls
                              # were sent and the remote requirements are too
                              # variable to trust reliable
-                             [[False, 'test_client_copy_file_double_payload']],
-                             indirect=True)
-    def test_client_copy_file_double_payload(self, wsman_conn):
-        client = self._get_client(wsman_conn)
-        client.wsman.max_payload_size = 113955
-
-        # data sent in 3 packets (2 * data + hash)
-        test_string = b"abcdefghijklmnopqrstuvwxyz" * 5000
-
-        temp_file, path = tempfile.mkstemp()
-        try:
-            os.write(temp_file, test_string)
-            actual = client.copy(path, "test_file")
-        finally:
-            os.close(temp_file)
-            os.remove(path)
-
-        # verify the returned object is the full path
-        assert actual == u"C:\\Users\\vagrant\\test_file"
-
-    @pytest.mark.parametrize('wsman_conn',
-                             # checks as to whether the correct number of calls
-                             # were sent and the remote requirements are too
-                             # variable to trust reliable
-                             [[False, 'test_client_copy_file_quad_payload']],
-                             indirect=True)
-    def test_client_copy_file_quad_payload(self, wsman_conn, monkeypatch):
-        # in a mocked context the calculated size differs on a few variables
-        # we will mock out that call and return the ones used in our existing
-        # responses
-        mock_calc = MagicMock()
-        mock_calc.side_effect = [113955, 382750]
-        monkeypatch.setattr(WSMan, "_calc_envelope_size", mock_calc)
-
-        client = self._get_client(wsman_conn)
-
-        # data sent in 3 packets (get size + data + hash)
-        test_string = b"abcdefghijklmnopqrstuvwxyz" * 10000
-
-        temp_file, path = tempfile.mkstemp()
-        try:
-            os.write(temp_file, test_string)
-            actual = client.copy(path, "test_file")
-        finally:
-            os.close(temp_file)
-            os.remove(path)
-
-        # verify the returned object is the full path
-        assert actual == u"C:\\Users\\vagrant\\test_file"
-
-    @pytest.mark.parametrize('wsman_conn',
-                             # checks as to whether the correct number of calls
-                             # were sent and the remote requirements are too
-                             # variable to trust reliable
                              [[False, 'test_client_copy_file_really_large']],
                              indirect=True)
     def test_client_copy_file_really_large(self, wsman_conn, monkeypatch):
@@ -143,12 +94,11 @@ class TestClient(object):
         # we will mock out that call and return the ones used in our existing
         # responses
         mock_calc = MagicMock()
-        mock_calc.side_effect = [113955, 382750]
+        mock_calc.return_value = 113955
         monkeypatch.setattr(WSMan, "_calc_envelope_size", mock_calc)
 
         client = self._get_client(wsman_conn)
 
-        # data sent in 4 packets (get size + 2 * data + hash)
         test_string = b"abcdefghijklmnopqrstuvwxyz" * 20000
 
         temp_file, path = tempfile.mkstemp()
@@ -160,7 +110,7 @@ class TestClient(object):
             os.remove(path)
 
         # verify the returned object is the full path
-        assert actual == u"C:\\Users\\vagrant\\test_file"
+        assert actual == u"C:\\Users\\vagrant\\Documents\\test_file"
 
     @pytest.mark.parametrize('wsman_conn',
                              [[True, 'test_client_copy_file_failure']],
@@ -182,10 +132,29 @@ class TestClient(object):
             with pytest.raises(WinRMError) as err:
                 actual = client.copy(path, "test_file")
             expected_err = \
-                "Failed to copy file: Transport failure, hash mistmatch\r\n" \
+                "Failed to copy file: Transport failure, hash mismatch\r\n" \
                 "Actual: 32d10c7b8cf96570ca04ce37f2a19d84240d3a89\r\n" \
                 "Expected: c3499c2729730a7f807efb8676a92dcb6f8a3f8f"
             assert expected_err in str(err.value)
+        finally:
+            os.close(temp_file)
+            os.remove(path)
+
+    @pytest.mark.parametrize('wsman_conn',
+                             # Requires a specially crafted response to be returned
+                             [[False, 'test_client_copy_file_warning']],
+                             indirect=True)
+    def test_client_copy_file_warning(self, wsman_conn):
+        client = self._get_client(wsman_conn)
+        test_string = b"abcdefghijklmnopqrstuvwxyz"
+
+        temp_file, path = tempfile.mkstemp()
+        try:
+            os.write(temp_file, test_string)
+            with pytest.warns(Warning, match="Failed to disable MaximumAllowedMemory input size: You cannot call a "
+                                             "method on a null-valued expression"):
+                actual = client.copy(path, "test_file")
+            client.execute_cmd("powershell Remove-Item -Path '%s'" % actual)
         finally:
             os.close(temp_file)
             os.remove(path)
@@ -207,6 +176,26 @@ class TestClient(object):
         assert actual_args[2] == 0
 
     @pytest.mark.parametrize('wsman_conn',
+                             [[True, 'test_client_execute_cmd_environment']], indirect=True)
+    def test_client_execute_cmd_environment(self, wsman_conn):
+        client = self._get_client(wsman_conn)
+
+        env = OrderedDict([
+            ('string', 'string value'),
+            ('int', 1234),
+            ('bool', True),
+            ('double_quote', 'double " quote'),
+            ('single_quote', "single ' quote"),
+            ('hyphen - var', 'abc @ 123'),
+        ])
+        actual = client.execute_cmd("set", environment=env)
+        actual_environment = actual[0].splitlines()
+        for env_key, env_value in env.items():
+            actual_env = next((e for e in actual_environment if e.startswith(env_key)), None)
+            assert actual_env is not None
+            assert actual_env == "%s=%s" % (env_key, env_value)
+
+    @pytest.mark.parametrize('wsman_conn',
                              [[True, 'test_client_execute_ps']], indirect=True)
     def test_client_execute_ps(self, wsman_conn):
         client = self._get_client(wsman_conn)
@@ -221,6 +210,27 @@ class TestClient(object):
         assert actual[0] == expected_stdout
         assert isinstance(actual[1], PSDataStreams)
         assert actual[2] is False
+
+    @pytest.mark.parametrize('wsman_conn',
+                             [[True, 'test_client_execute_ps_environment']], indirect=True)
+    def test_client_execute_ps_environment(self, wsman_conn):
+        client = self._get_client(wsman_conn)
+
+        env = OrderedDict([
+            ('string', 'string value'),
+            ('int', 1234),
+            ('bool', True),
+            ('double_quote', 'double " quote'),
+            ('single_quote', "single ' quote"),
+            ('hyphen - var', 'abc @ 123'),
+            ('_-(){}[]<>*+-/\\?"''!@#$%^&|;:i,.`~0', '_-(){}[]<>*+-/\\?"''!@#$%^&|;:i,.`~0'),
+        ])
+        actual = client.execute_ps("cmd.exe /c set", environment=env)
+        actual_environment = actual[0].splitlines()
+        for env_key, env_value in env.items():
+            actual_env = next((e for e in actual_environment if e.startswith(env_key)), None)
+            assert actual_env is not None
+            assert actual_env == "%s=%s" % (env_key, env_value)
 
     @pytest.mark.parametrize('wsman_conn',
                              [[True, 'test_client_execute_ps_failure']],
@@ -243,14 +253,7 @@ class TestClient(object):
                              # means we don't need to create files on the
                              # remote side
                              [[False, 'test_client_fetch_file']], indirect=True)
-    def test_client_fetch_file(self, wsman_conn, monkeypatch):
-        # in a mocked context the calculated size differs on a few variables
-        # we will mock out that call and return the ones used in our existing
-        # responses
-        mock_calc = MagicMock()
-        mock_calc.side_effect = [113955, 382750]
-        monkeypatch.setattr(WSMan, "_calc_envelope_size", mock_calc)
-
+    def test_client_fetch_file(self, wsman_conn):
         client = self._get_client(wsman_conn)
 
         # file was created with
@@ -280,14 +283,7 @@ class TestClient(object):
     @pytest.mark.parametrize('wsman_conn',
                              [[True, 'test_client_fetch_file_fail_dir']],
                              indirect=True)
-    def test_client_fetch_file_fail_dir(self, wsman_conn, monkeypatch):
-        # in a mocked context the calculated size differs on a few variables
-        # we will mock out that call and return the ones used in our existing
-        # responses
-        mock_calc = MagicMock()
-        mock_calc.side_effect = [113955, 382750]
-        monkeypatch.setattr(WSMan, "_calc_envelope_size", mock_calc)
-
+    def test_client_fetch_file_fail_dir(self, wsman_conn):
         client = self._get_client(wsman_conn)
         with pytest.raises(WinRMError) as err:
             client.fetch("C:\\Windows", "")
@@ -298,15 +294,7 @@ class TestClient(object):
     @pytest.mark.parametrize('wsman_conn',
                              [[True, 'test_client_fetch_file_fail_missing']],
                              indirect=True)
-    def test_client_fetch_file_fail_missing(self, wsman_conn,
-                                            monkeypatch):
-        # in a mocked context the calculated size differs on a few variables
-        # we will mock out that call and return the ones used in our existing
-        # responses
-        mock_calc = MagicMock()
-        mock_calc.side_effect = [113955, 382750]
-        monkeypatch.setattr(WSMan, "_calc_envelope_size", mock_calc)
-
+    def test_client_fetch_file_fail_missing(self, wsman_conn):
         client = self._get_client(wsman_conn)
         with pytest.raises(WinRMError) as err:
             client.fetch("C:\\fakefile.txt", "")
@@ -337,3 +325,37 @@ class TestClient(object):
             "Failed to fetch file C:\\temp\\file.txt, hash mismatch\n" \
             "Source: eec729d9a0fa275513bc44a4cb8d4ee973b81e1a\n" \
             "Fetched: c3499c2729730a7f807efb8676a92dcb6f8a3f8f"
+
+    def test_sanitise_clixml_with_error(self):
+        clixml_path = os.path.join(os.path.dirname(__file__), 'data', 'test_sanitise_clixml_with_error.xml')
+        with open(clixml_path, 'r') as fd:
+            clixml = fd.read()
+
+        expected = "fake : The term 'fake' is not recognized as the name of a cmdlet, function, script file, or operable program. Check \r\n" \
+                   "the spelling of the name, or if a path was included, verify that the path is correct and try again.\r\n" \
+                   "At line:1 char:1\r\n" \
+                   "+ fake cmdlet\r\n" \
+                   "+ ~~~~\r\n" \
+                   "    + CategoryInfo          : ObjectNotFound: (fake:String) [], CommandNotFoundException\r\n" \
+                   "    + FullyQualifiedErrorId : CommandNotFoundException\r\n" \
+                   " \r\n"
+
+        actual = Client.sanitise_clixml(clixml)
+        assert actual == expected
+
+    def test_sanitise_clixml_with_no_errors(self):
+        clixml_path = os.path.join(os.path.dirname(__file__), 'data', 'test_sanitise_clixml_with_no_errors.xml')
+        with open(clixml_path, 'r') as fd:
+            clixml = fd.read()
+
+        expected = ""
+
+        actual = Client.sanitise_clixml(clixml)
+        assert actual == expected
+
+    def test_sanitise_clixml_not_clixml(self):
+        clixml = "stderr line"
+        expected = clixml
+
+        actual = Client.sanitise_clixml(clixml)
+        assert actual == expected